requirements.txt
alembic==1.7.4
antiorm==1.2.1
blinker==1.4
Bootstrap-Flask==1.8.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
db==0.1.1
Flask==2.0.2
Flask-Ext==0.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Mpesa==1.2.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
fpdf==1.7.2
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
Mako==1.1.5
MarkupSafe==2.0.1
passlib==1.7.4
phonenumbers==8.12.35
Pillow==8.4.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
twilio==7.1.0
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==2.3.3
